+++ commons/src/main/scala/commons.scala
@@ -187,6 +187,9 @@ object domain {
   import zio.json.internal.Write
   import java.time.{LocalDateTime, ZonedDateTime}
 
+  val Presto      = ApplicationCode.of("presto")
+  val RootAccount = AccountId.of(1)
+
   trait HasEmail {
     def email: Email
   }