+++ server/src/main/scala/billing.scala
@@ -4,7 +4,6 @@ package morbid
 object billing {
 
   import zio.*
-  import morbid.commands.ReportUsersByAccount
   import morbid.domain.raw.RawAccount
   import morbid.types.{AccountCode, ApplicationCode}
   import morbid.repo.Repo
@@ -18,6 +17,6 @@ object billing {
   }
 
   case class SimpleBilling(repo: Repo) extends Billing {
-    override def usersByAccount(app: ApplicationCode): Task[Map[RawAccount, Int]] = repo.exec(ReportUsersByAccount(app))
+    override def usersByAccount(app: ApplicationCode): Task[Map[RawAccount, Int]] = ??? // repo.exec(ReportUsersByAccount(app))
   }
 }